# fixes #16749 - improve NIC update performance during fact imports