@lzap 2016-10-04
Thanks @domcleal! Merged.